logeverything-dashboard/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logeverything-dashboard"
version = "0.1.0"
description = "Real-time monitoring dashboard for LogEverything"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [{name = "LogEverything Team"}]
keywords = [
    "logging",
    "monitoring",
    "dashboard",
    "logeverything",
    "observability",
    "fastapi",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
    "Framework :: AsyncIO",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "libsass>=0.22.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.19.0",
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "websockets>=12.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "build>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/logeverything/logeverything"
Documentation = "https://logeverything.readthedocs.io/en/latest/user-guide/dashboard.html"
Repository = "https://github.com/logeverything/logeverything-dashboard"
Issues = "https://github.com/logeverything/logeverything-dashboard/issues"

[project.scripts]
logeverything-dashboard = "dashboard.cli:main"

[tool.setuptools.packages.find]
include = ["dashboard*"]

[tool.setuptools.package-data]
dashboard = [
    "templates/**/*",
    "static/**/*",
    "py.typed",
]
